Fixed wrong display size of really big packs
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@1886 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/panels/PacksModel.java
@@ -311,7 +311,7 @@ class PacksModel extends AbstractTableModel
             }
 
         case 2:
-            return Pack.toByteUnitsString((int) pack.nbytes);
+            return Pack.toByteUnitsString(pack.nbytes);
 
         default:
             return null;
